Started bin drop-off optimizations
Fixed the code going to the bin drop. Started trying to make the bin
drop in the same place each time. New strategy is to use the outrigger
to follow the black line connected to the habitat until it detects the
line perpendicular to the debris. The outrigger motor has been switched
with the bin grabber servo to allow for greater movement. In order to
make the line follow consistent the outrigger will have to rotate until
it hits the habitat-connected line and then fix itself to the required
position while linefollowing.