Minor optimizations
Today we changed the way that the robot finds the bin when dumping the
first pile. It now goes until it sees a black line and then slow drives
with et. Also, we changed the way that the robot turns when backing up
from the bin (after dumping first pile). Additionally we made sure that
the robot will not hit the coupler after dumping bin. The line follow
before squaring up for the bin drop off should be shortened and/or
slowed down increase its effectiveness. The sensor test still needs to
be added.