Clone code and prime fixes
Today we changed multiple servo values on clone to be able to mirror the
prime code. Notably it was the turn for the north pile and the back up
to grab the bin. The bin back up may need to be further adjusted.
Clone's turn to go into the center needs to be made more reliable. This
may have been fixed, but it is relatively untested. The amount of
rotation to grab the block in the valley has also been adjusted. It
seems to have fixed the problem of missing the block, but has only been
attempted once. This is in the grabCube code. The amount of line follow
before the square up on the pipe may need to be adjusted.